Update execution_order, loop_output samples to v2 pipelines
