+++ samples/core/execution_order/execution_order_test.py
@@ -12,12 +12,12 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import kfp.deprecated as kfp
+import kfp
 from kfp.samples.test.utils import TestCase, relative_path, run_pipeline_func
 
 run_pipeline_func([
     TestCase(
         pipeline_file=relative_path(__file__, 'execution_order.py'),
-        mode=kfp.dsl.PipelineExecutionMode.V1_LEGACY,
+        mode=kfp.dsl.PipelineExecutionMode.V2_ENGINE,
     ),
 ])
