+++ samples/core/loop_output/loop_output_test.py
@@ -15,11 +15,10 @@
 from __future__ import annotations
 
 import unittest
-import kfp.deprecated as kfp
+import kfp
 import kfp_server_api
-from ml_metadata.proto import Execution
-from .loop_output import my_pipeline
-from .loop_output_v2 import my_pipeline as my_pipeline_v2
+from ml_metadata.proto.metadata_store_pb2 import Execution
+from loop_output import my_pipeline
 from kfp.samples.test.utils import KfpTask, run_pipeline_func, TestCase
 
 
@@ -76,12 +75,8 @@ def verify(t: unittest.TestCase, run: kfp_server_api.ApiRun,
 
 run_pipeline_func([
     TestCase(
-        pipeline_func=my_pipeline_v2,
+        pipeline_func=my_pipeline,
         mode=kfp.dsl.PipelineExecutionMode.V2_ENGINE,
         verify_func=verify,
-    ),
-    TestCase(
-        pipeline_func=my_pipeline,
-        mode=kfp.dsl.PipelineExecutionMode.V1_LEGACY,
     ),
 ])
